+++ packages/@phased/phase/app-client.js
@@ -1,0 +1,3 @@
+import app from '../../../resources/js/app';
+
+app.$mount('#app');